+++ src/main/java/org/lxdproject/lxd/friend/service/FriendService.java
@@ -6,11 +6,12 @@ import lombok.extern.slf4j.Slf4j;
 import org.lxdproject.lxd.apiPayload.code.exception.handler.FriendHandler;
 import org.lxdproject.lxd.apiPayload.code.status.ErrorStatus;
 
+import org.lxdproject.lxd.authz.guard.FriendGuard;
+import org.lxdproject.lxd.authz.guard.MemberGuard;
 import org.lxdproject.lxd.common.dto.PageDTO;
 import org.lxdproject.lxd.config.security.SecurityUtil;
 import org.lxdproject.lxd.friend.dto.*;
 import org.lxdproject.lxd.friend.entity.FriendRequest;
-import org.lxdproject.lxd.infra.redis.RedisKeyPrefix;
 import org.lxdproject.lxd.infra.redis.RedisService;
 import org.lxdproject.lxd.member.entity.Member;
 import org.lxdproject.lxd.friend.entity.enums.FriendRequestStatus;
@@ -22,14 +23,12 @@ import org.lxdproject.lxd.notification.entity.enums.NotificationType;
 import org.lxdproject.lxd.notification.entity.enums.TargetType;
 import org.lxdproject.lxd.notification.repository.NotificationRepository;
 import org.lxdproject.lxd.notification.service.NotificationService;
-import org.lxdproject.lxd.notification.service.SseEmitterService;
+import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.transaction.support.TransactionSynchronization;
-import org.springframework.transaction.support.TransactionSynchronizationManager;
 
 import java.util.Collections;
 import java.util.List;
@@ -44,10 +43,13 @@ public class FriendService {
     private final MemberRepository memberRepository;
     private final FriendRequestRepository friendRequestRepository;
 
+    private final ApplicationEventPublisher eventPublisher;
     private final NotificationService notificationService;
-    private final SseEmitterService sseEmitterService;
     private final NotificationRepository notificationRepository;
     private final RedisService redisService;
+
+    private final FriendGuard friendGuard;
+    private final MemberGuard memberGuard;
 
     @Transactional(readOnly = true)
     public FriendListResponseDTO getFriendList(Long memberId, int page, int size) {
@@ -77,6 +79,7 @@ public class FriendService {
         return new FriendListResponseDTO(totalRequests, pageDTO);
     }
 
+    @Transactional
     public void sendFriendRequest(Long requesterId, FriendRequestCreateRequestDTO requestDto) {
         Long receiverId = requestDto.getReceiverId();
 
@@ -86,9 +89,14 @@ public class FriendService {
 
         Member requester = memberRepository.findById(requesterId)
                 .orElseThrow(() -> new FriendHandler(ErrorStatus.MEMBER_NOT_FOUND));
+        memberGuard.checkOwnerIsNotDeleted(requester);
 
         Member receiver = memberRepository.findById(receiverId)
                 .orElseThrow(() -> new FriendHandler(ErrorStatus.MEMBER_NOT_FOUND));
+        memberGuard.checkOwnerIsNotDeleted(receiver);
+
+        // 친구 요청이 가능한 지에 대하 인가 검사
+        friendGuard.validateBeforeRequestAction(requester, receiver);
 
         boolean alreadyRequested = friendRequestRepository.existsByRequesterAndReceiverAndStatus(
                 requester, receiver, FriendRequestStatus.PENDING);
@@ -114,13 +122,14 @@ public class FriendService {
 
         NotificationRequestDTO dto = NotificationRequestDTO.builder()
                 .receiverId(receiver.getId())
+                .senderId(requester.getId())
                 .notificationType(NotificationType.FRIEND_REQUEST)
                 .targetType(TargetType.MEMBER)
                 .targetId(requester.getId())
                 .redirectUrl("/members/" + requester.getId())
                 .build();
 
-        notificationService.saveAndPublishNotification(dto);
+        notificationService.createAndPublish(dto);
     }
 
     @Transactional
@@ -134,54 +143,54 @@ public class FriendService {
             throw new FriendHandler(ErrorStatus.FRIEND_REQUEST_NOT_PENDING);
         }
 
-        // 친구 관계 양방향 저장
         Member requester = request.getRequester();
+        memberGuard.checkOwnerIsNotDeleted(requester);
         Member receiver = request.getReceiver();
+        memberGuard.checkOwnerIsNotDeleted(receiver);
+
+        // 친구 요청 수락에 대한 인가 검사
+        friendGuard.validateBeforeRequestAction(requester, receiver);
+
+        // 친구 관계 양방향 저장
         friendRepository.saveFriendship(requester, receiver);
 
-        // 친구 요청 알림 삭제
-        long deleted = notificationRepository.deleteFriendRequestNotification(receiver.getId(), requester.getId());
-        if (deleted > 0) {
-            notificationRepository.flush();
-
-            TransactionSynchronizationManager.registerSynchronization(new TransactionSynchronization() {
-                @Override
-                public void afterCommit() {
-                    sseEmitterService.sendNotificationDeleted(
-                            receiverId,
-                            NotificationType.FRIEND_REQUEST,
-                            TargetType.MEMBER,
-                            requesterId
-                    );
-                }
-            });
-        }
+        // 기존 친구 요청 알림 삭제
+        notificationService.deleteAndPublish(
+                receiverId,
+                requesterId,
+                NotificationType.FRIEND_REQUEST,
+                TargetType.MEMBER,
+                requesterId
+        );
 
         // 친구 요청 엔티티 삭제
         friendRequestRepository.delete(request);
 
-        // 친구 요청자에게 친구 수락됨 알림 보내기
+        // 친구 요청자에게 친구 수락됨 알림 이벤트 발행
         NotificationRequestDTO dto = NotificationRequestDTO.builder()
-                .receiverId(requester.getId()) // 친구 요청 보낸 사람에게 알림 전송
+                .receiverId(requester.getId()) // 요청 보낸 사람에게 알림
+                .senderId(receiver.getId()) // 요청 수락한 사람
                 .notificationType(NotificationType.FRIEND_ACCEPTED)
                 .targetType(TargetType.MEMBER)
                 .targetId(receiver.getId()) // 친구 요청 수락한 사람
                 .redirectUrl("/members/" + receiver.getId())
                 .build();
 
-        notificationService.saveAndPublishNotification(dto);
+        notificationService.createAndPublish(dto);
     }
 
+    @Transactional
     public void deleteFriend(Long currentMemberId, Long friendId) {
         Member current = memberRepository.findById(currentMemberId)
                 .orElseThrow(() -> new FriendHandler(ErrorStatus.MEMBER_NOT_FOUND));
+        memberGuard.checkOwnerIsNotDeleted(current);
+
         Member target = memberRepository.findById(friendId)
                 .orElseThrow(() -> new FriendHandler(ErrorStatus.MEMBER_NOT_FOUND));
+        memberGuard.checkOwnerIsNotDeleted(target);
 
-        boolean exists = friendRepository.areFriends(currentMemberId, friendId);
-        if (!exists) {
-            throw new FriendHandler(ErrorStatus.NOT_FRIEND);
-        }
+        // 친구 삭제에 대한 인가 검사
+        friendGuard.validateBeforeManageAction(target, current);
 
         friendRepository.deleteFriendship(current, target);
     }
@@ -236,41 +245,34 @@ public class FriendService {
         return friendRequestRepository.countByReceiverAndStatus(member, FriendRequestStatus.PENDING);
     }
 
-    private FriendResponseDTO mapToDto(Member member) {
-        return new FriendResponseDTO(
-                member.getId(),
-                member.getUsername(),
-                member.getNickname(),
-                member.getProfileImg()
-        );
-    }
-
     @Transactional
     public void refuseFriendRequest(FriendRequestRefuseRequestDTO requestDto) {
         Long receiverId = SecurityUtil.getCurrentMemberId();
         Long requesterId = requestDto.getRequesterId();
 
+        Member receiver = findMemberById(receiverId);
+        // 탈퇴한 사용자인지 검사
+        memberGuard.checkOwnerIsNotDeleted(receiver);
+
+        Member requester = findMemberById(requesterId);
+        // 탈퇴한 사용자인지 검사
+        memberGuard.checkOwnerIsNotDeleted(requester);
+
+        // 친구 요청 거절에 대하 인가 검사
+        friendGuard.validateBeforeRequestAction(requester, receiver);
+
         FriendRequest request = friendRequestRepository
                 .findByRequesterIdAndReceiverIdAndStatus(requesterId, receiverId, FriendRequestStatus.PENDING)
                 .orElseThrow(() -> new FriendHandler(ErrorStatus.FRIEND_REQUEST_NOT_FOUND));
 
-        // 친구 요청 알림 삭제
-        long deleted = notificationRepository.deleteFriendRequestNotification(receiverId, requesterId);
-        if (deleted > 0) {
-            notificationRepository.flush();
-
-            TransactionSynchronizationManager.registerSynchronization(new TransactionSynchronization() {
-                @Override
-                public void afterCommit() {
-                    sseEmitterService.sendNotificationDeleted(
-                            receiverId,
-                            NotificationType.FRIEND_REQUEST,
-                            TargetType.MEMBER,
-                            requesterId
-                    );
-                }
-            });
-        }
+        // 기존 찬구 요청 알림 삭제
+        notificationService.deleteAndPublish(
+                receiverId,
+                requesterId,
+                NotificationType.FRIEND_REQUEST,
+                TargetType.MEMBER,
+                requesterId
+        );
 
         // 친구 요청 엔티티 삭제
         friendRequestRepository.delete(request);
@@ -281,26 +283,29 @@ public class FriendService {
         Long requesterId = SecurityUtil.getCurrentMemberId();
         Long receiverId = requestDto.getReceiverId();
 
+        Member receiver = findMemberById(receiverId);
+        // 탈퇴한 사용자인지 검사
+        memberGuard.checkOwnerIsNotDeleted(receiver);
+
+        Member requester = findMemberById(requesterId);
+        // 탈퇴한 사용자인지 검사
+        memberGuard.checkOwnerIsNotDeleted(requester);
+
+        // 친구 요청 취소에 대하 인가 검사
+        friendGuard.validateBeforeRequestAction(requester, receiver);
+
         FriendRequest request = friendRequestRepository
                 .findByRequesterIdAndReceiverIdAndStatus(requesterId, receiverId, FriendRequestStatus.PENDING)
                 .orElseThrow(() -> new FriendHandler(ErrorStatus.FRIEND_REQUEST_NOT_FOUND));
 
-        long deleted = notificationRepository.deleteFriendRequestNotification(receiverId, requesterId);
-        if (deleted > 0) {
-            notificationRepository.flush();
-
-            TransactionSynchronizationManager.registerSynchronization(new TransactionSynchronization() {
-                @Override
-                public void afterCommit() {
-                    sseEmitterService.sendNotificationDeleted(
-                            receiverId,
-                            NotificationType.FRIEND_REQUEST,
-                            TargetType.MEMBER,
-                            requesterId
-                    );
-                }
-            });
-        }
+        // 기존 친구 요청 알림 삭제
+        notificationService.deleteAndPublish(
+                receiverId,
+                requesterId,
+                NotificationType.FRIEND_REQUEST,
+                TargetType.MEMBER,
+                requesterId
+        );
 
         friendRequestRepository.delete(request);
     }
